Pulls from the database now!

+++ BitemporalVisualization/BrushProvider.cs
@@ -12,26 +12,26 @@ namespace BitemporalVisualization
     public class BrushProvider
     {
         private List<Gradient> gradients; 
-        private Dictionary<int, Gradient> gradientMappings;
-        private int highlightedTransaction;
+        private Dictionary<long, Gradient> gradientMappings;
+        private long highlightedTransaction;
 
-        public BrushProvider(int tId)
+        public BrushProvider(long tId)
         {
             highlightedTransaction = tId;
             gradients = new List<Gradient>();
-            gradientMappings = new Dictionary<int, Gradient>();
+            gradientMappings = new Dictionary<long, Gradient>();
             gradients.Add(new Gradient(Color.FromArgb(214, 54, 201), Color.FromArgb(173, 0, 159)));
             gradients.Add(new Gradient(Color.FromArgb(255, 126, 54), Color.FromArgb(255, 83, 0)));
             gradients.Add(new Gradient(Color.FromArgb(53, 213, 157), Color.FromArgb(0, 171, 111)));
             gradients.Add(new Gradient(Color.FromArgb(214, 250, 63), Color.FromArgb(157, 184, 46)));
         }
 
-        public Brush GetBrush(int transactionId, int revisionId, DateTime recordFrom, DateTime recordTo, DateTime validFrom, DateTime validTo, Rectangle bounds)
+        public Brush GetBrush(long transactionId, long revisionId, DateTime recordFrom, DateTime recordTo, DateTime validFrom, DateTime validTo, Rectangle bounds)
         {
             if (bounds.Width == 0 || bounds.Height == 0)
                 return new SolidBrush(Color.Black);
             if (!gradientMappings.ContainsKey(revisionId))
-                gradientMappings[revisionId] = gradients[revisionId%gradients.Count];
+                gradientMappings[revisionId] = gradients[(int)revisionId%gradients.Count];
 
             LinearGradientBrush br = new LinearGradientBrush(bounds, Color.Black, Color.Black, 0, false);
             ColorBlend cb = new ColorBlend();
